First full upload of assignment one!
There are still some pieces I'd like to come back and fix up, namely the fact that the mockups/wireframes show that there is no 'Read this blog' link when you're viewing on a mobile device on the individual blog pages, which seems a little odd to me. Some of the layout could be improved a lil bit too but I'm much happier that things are looking a bit better. I plan to revisit this and tidy it up a bit more after assignment two but hey, it's functional for now (I think :) )